refactor CheckChatCompletionsWithURLAndHeaders
