+++ test/e2e/utils/chat.go
@@ -71,7 +71,9 @@ func CheckChatCompletionsWithHeaders(t *testing.T, modelName string, messages []
 	return CheckChatCompletionsWithURLAndHeaders(t, DefaultRouterURL, modelName, messages, headers)
 }
 
-func CheckChatCompletionsWithURLAndHeaders(t *testing.T, url string, modelName string, messages []ChatMessage, headers map[string]string) *ChatCompletionsResponse {
+// SendChatRequestWithRetry sends a chat completions request with retry logic but without assertions.
+// It returns the final response regardless of status code.
+func SendChatRequestWithRetry(t *testing.T, url string, modelName string, messages []ChatMessage, headers map[string]string) *ChatCompletionsResponse {
 	requestBody := ChatCompletionsRequest{
 		Model:    modelName,
 		Messages: messages,
@@ -145,12 +147,9 @@ func CheckChatCompletionsWithURLAndHeaders(t *testing.T, url string, modelName s
 			continue
 		}
 
-		// Last attempt, verify response
+		// Last attempt - log the response but don't assert success
 		t.Logf("Chat response status: %d", resp.StatusCode)
 		t.Logf("Chat response: %s", responseStr)
-		assert.Equal(t, http.StatusOK, resp.StatusCode, "Expected HTTP 200 status code")
-		assert.NotEmpty(t, responseStr, "Chat response is empty")
-		assert.NotContains(t, responseStr, "error", "Chat response contains error")
 		break
 	}
 
@@ -158,6 +157,17 @@ func CheckChatCompletionsWithURLAndHeaders(t *testing.T, url string, modelName s
 		StatusCode: resp.StatusCode,
 		Body:       responseStr,
 	}
+}
+
+func CheckChatCompletionsWithURLAndHeaders(t *testing.T, url string, modelName string, messages []ChatMessage, headers map[string]string) *ChatCompletionsResponse {
+	resp := SendChatRequestWithRetry(t, url, modelName, messages, headers)
+
+	// Assert successful response
+	assert.Equal(t, http.StatusOK, resp.StatusCode, "Expected HTTP 200 status code")
+	assert.NotEmpty(t, resp.Body, "Chat response is empty")
+	assert.NotContains(t, resp.Body, "error", "Chat response contains error")
+
+	return resp
 }
 
 // containsError checks if the response string contains error indicators
@@ -204,69 +214,4 @@ func SendChatRequestWithURL(t *testing.T, url string, modelName string, messages
 	require.NoError(t, err, "Failed to send HTTP request")
 
 	return resp
-}
-
-// LoadLoRAAdapter loads a LoRA adapter directly on the LLM-Mock pod by sending a request to /v1/load_lora_adapter
-// The request is sent directly to the specified pod URL (e.g., http://127.0.0.1:9000/v1/load_lora_adapter)
-// Note: This should NOT be sent through the router, as /v1/load_lora_adapter is a management endpoint
-func LoadLoRAAdapter(t *testing.T, podURL string, loraName string, loraPath string) {
-	loadURL := strings.TrimSuffix(podURL, "/") + "/v1/load_lora_adapter"
-
-	requestBody := map[string]interface{}{
-		"lora_name": loraName,
-		"lora_path": loraPath,
-	}
-
-	jsonData, err := json.Marshal(requestBody)
-	require.NoError(t, err, "Failed to marshal load LoRA adapter request body")
-
-	client := &http.Client{
-		Timeout: 30 * time.Second,
-	}
-
-	req, err := http.NewRequest("POST", loadURL, bytes.NewBuffer(jsonData))
-	require.NoError(t, err, "Failed to create HTTP request")
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := client.Do(req)
-	require.NoError(t, err, "Failed to send HTTP request")
-	defer resp.Body.Close()
-
-	responseBody, err := io.ReadAll(resp.Body)
-	require.NoError(t, err, "Failed to read response body")
-
-	assert.Equal(t, http.StatusOK, resp.StatusCode, "Expected HTTP 200 for loading LoRA adapter")
-	t.Logf("Successfully loaded LoRA adapter %s: %s", loraName, string(responseBody))
-}
-
-// UnloadLoRAAdapter unloads a LoRA adapter directly on the LLM-Mock pod by sending a request to /v1/unload_lora_adapter
-// The request is sent directly to the specified pod URL (e.g., http://127.0.0.1:9000/v1/unload_lora_adapter)
-// Note: This should NOT be sent through the router, as /v1/unload_lora_adapter is a management endpoint
-func UnloadLoRAAdapter(t *testing.T, podURL string, loraName string) {
-	unloadURL := strings.TrimSuffix(podURL, "/") + "/v1/unload_lora_adapter"
-
-	requestBody := map[string]interface{}{
-		"lora_name": loraName,
-	}
-
-	jsonData, err := json.Marshal(requestBody)
-	require.NoError(t, err, "Failed to marshal unload LoRA adapter request body")
-
-	client := &http.Client{
-		Timeout: 30 * time.Second,
-	}
-
-	req, err := http.NewRequest("POST", unloadURL, bytes.NewBuffer(jsonData))
-	require.NoError(t, err, "Failed to create HTTP request")
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := client.Do(req)
-	require.NoError(t, err, "Failed to send HTTP request")
-	defer resp.Body.Close()
-
-	responseBody, err := io.ReadAll(resp.Body)
-	require.NoError(t, err, "Failed to read response body")
-
-	assert.Equal(t, http.StatusOK, resp.StatusCode, "Expected HTTP 200 for unloading LoRA adapter")
-	t.Logf("Successfully unloaded LoRA adapter %s: %s", loraName, string(responseBody))
 }
